code/requirements.txt
asn1crypto==0.24.0
celery==4.1.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
channels==2.1.2
channels-redis==2.2.1
cryptography==2.3
docker==3.4.1
Django==2.0.7
django-authtools==1.6.0
django-dotenv==1.4.2
django-extensions==2.1.0
djangorestframework==3.8.2
social-auth-app-django==2.1.0
django-webpack-loader==0.6.0
hvac[parser]==0.6.2
jinja2==2.10
paramiko==2.4.1
PyGithub==1.40
psycopg2-binary==2.7.5
kubernetes==6.0.0
idna==2.7
pycparser==2.18
pytz==2018.5
requests==2.19.1
passlib==1.7.1
six==1.11.0
stripe==2.0.3
urllib3==1.23
websocket-client==0.48.0